+++ src/FluentlyHttpClient/FluentHttpRequest.cs
@@ -66,18 +66,18 @@ namespace FluentlyHttpClient
 		/// Formatters to be used for content negotiation for "Accept" and also sending formats. e.g. (JSON, XML)
 		/// </summary>
 		[Obsolete("This was added to be passed down to the middleware. Instead in middleware use FluentHttpMiddlewareClientContext.Formatters.")]
-		public MediaTypeFormatterCollection Formatters { get; set; }
+		public MediaTypeFormatterCollection Formatters { get; set; } // deprecated: remove
 
 		/// <summary>
 		/// Initializes a new instance.
 		/// </summary>
-		public FluentHttpRequest(FluentHttpRequestBuilder builder, HttpRequestMessage message, IDictionary<object, object> items = null)
+		public FluentHttpRequest(FluentHttpRequestBuilder builder, HttpRequestMessage message, IDictionary<object, object> items = null) // todo: is items really needed?
 		{
 			Message = message;
 			Builder = builder;
-			Items = items == null
-				? new Dictionary<object, object>()
-				: new Dictionary<object, object>(items);
+			Items = new Dictionary<object, object>(
+				items ?? builder.Items ?? new Dictionary<object, object>()
+			);
 		}
 
 		/// <summary>